requirements.txt
pdfplumber
groq
yfinance
pandas
numpy
feedparser
transformers
torch
fastapi
uvicorn
python-dotenv
streamlit
plotly
requests